Send cronjob mail to the user who signed up a locker by default
Previously, we set up crontabs to send mail to lockername@mit.edu.
This was fine when scripts didn't support non-user lockers.  It's
not necessarily a fine assumption in the case of group lockers that
the address is appropriate (or even exists).  We could send mail
to the scripts user, but that only works if the user is signed up
for mail_scripts.  Instead, we'll default it to the user who signed
the locker up for cron_scripts (the same logic mail_scripts uses
for the initial procmailrc) and tell them to change it if it should
go elsewhere when they set up their crontab.

Also move the crontab template to bin instead of deploy so that it
can be under revision control.  Finally, correct a whitespace error
in the ASCII art (reported by kaduk).

Legacy-ID: 1521