Additional updates for Django 1.4/1.5: static files, admin interface, README
Here we provide proper support for static files on Scripts, see
	<https://docs.djangoproject.com/en/1.4/howto/static-files/>

This means we designate a path for static files (~/web_scripts/$name/static)
and refer to the static files by their full username.scripts.mit.edu url over
HTTPS. This should work even if the user deploys the app on a subdomain or a
non-MIT hostname.

At the same time, we fix the urls.py rule to enable the admin interface,
which broke with the new URL pattern syntax a while ago. 

Finally, we run collectstatic to ensure that the relevant files are copied over.

In addition, a README was added to the web_scripts directory to help avoid
confusion over "where're my files?".


Legacy-ID: 2441